Transforming and augmenting images
The train_test_split function now considers transformation functions for both the train and the test dataset. To apply the transformation function to each patch, an additional subclass CustomTransformedDataset is introduced. Note that the labels are augmented in line with the patches (e.g. if the patch is flipped also the label is flipped).